+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/StrengthStabilityPointConstructionSectionResultEntity.cs
@@ -37,8 +37,8 @@ namespace Application.Ringtoets.Storage.DbContext
     {
         public long StrengthStabilityPointConstructionSectionResultEntityId { get; set; }
         public long FailureMechanismSectionEntityId { get; set; }
-        public Nullable<decimal> LayerTwoA { get; set; }
-        public Nullable<decimal> LayerThree { get; set; }
+        public Nullable<double> LayerTwoA { get; set; }
+        public Nullable<double> LayerThree { get; set; }
     
         public virtual FailureMechanismSectionEntity FailureMechanismSectionEntity { get; set; }
     }
